ai_config/requirement.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.2.0
decouple==0.0.7
dill==0.3.8
Django==5.1.6
djangorestframework==3.15.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.28.1
idna==3.10
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.2
packaging==24.2
pandas==2.2.3
propcache==0.2.1
psutil==6.1.1
pyarrow==19.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
